+++ .vscode/settings.json
@@ -1,4 +1,13 @@
 {
     "python.linting.enabled": true,
-    "appMap.findingsEnabled": true
+    "appMap.findingsEnabled": true,
+    "python.testing.unittestArgs": [
+        "-v",
+        "-s",
+        "./test",
+        "-p",
+        "test_*.py"
+    ],
+    "python.testing.pytestEnabled": false,
+    "python.testing.unittestEnabled": true
 }
