+++ app/Main.py
@@ -1,0 +1,6 @@
+from src.tkresource.tkinter_wrapper import root
+from src.menu.mainmenu import open
+
+open()
+
+root.mainloop()
