+++ app/src/display/obstacle/basic/resource/picture.py
@@ -3,7 +3,7 @@ from PIL import ImageTk, Image
 class Picture:
     def __init__(self, folder, file):
         self.angle = 0
-        self.file = f'display/{folder}sprites/{file}'
+        self.file = f'app/src/display/{folder}sprites/{file}'
         self.image = Image.open(self.file)
 
     def display(self, displayer, x : int, y : int, anchor : str, tag : str):
