+++ app/src/display/obstacle/sprites/Stone.png
