+++ app/src/display/obstacle/stone.py
