+++ app/src/display/rover/sprites/Rover.png
