+++ app/src/menu/mainmenu.py
@@ -1,9 +1,9 @@
-from tkresource.tkinter_widgets import Widgets, widgets_remove
-from tkresource.tkinter_wrapper import root, root_width, root_height
-from menu.configuration.config import Configuration
+from ..tkresource.tkinter_widgets import Widgets, widgets_remove
+from ..tkresource.tkinter_wrapper import root, root_width, root_height
+from .configuration.config import Configuration
 
 def open():
-    from menu.level.level1 import Level1
+    from ..menu.level.level1 import Level1
     print('>>> Start mainmenu')
     widgets_remove(root)
     Widgets(root, 'START', 'center', root_width/2, root_height/2 - 50).sizedbutton(Level1, 15)
