--- src/Main.py
@@ -1,6 +1,0 @@
-from tkresource.tkinter_wrapper import root
-from menu.mainmenu import open
-
-open()
-
-root.mainloop()
